Move mysql stuff to dedicated class
